+++ packages/terra-doc-template/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Added a visible CSS layout
+* Removed terra-toggle-button as a dependency
 
 1.2.0 - (May 16, 2018)
 ------------------